Redundancy reduction & Spelling correction
I'm opening this pull request proposing, the removal of word 'of' that was previously written twice by mistake in the documentation. Also, there was a spelling mistake in how 'interpretted' was written. It has been corrected & changed to 'interpreted'. Please review it once for the corrections being proposed. I hope this helps to improve your repository.